New tests were added
