--- include/add.h
@@ -1,8 +1,0 @@
-#ifndef INCLUDE_ADD_H_
-#define INCLUDE_ADD_H_
-
-struct CNode {
-	int x, y;
-};
-char* ways(CNode m);
-#endif  // INCLUDE_ADD_H_
