+++ include/ways.h
@@ -1,0 +1,8 @@
+#ifndef INCLUDE_WAYS_H_
+#define INCLUDE_WAYS_H_
+
+struct CNode {
+    int x, y;
+};
+char* ways(CNode m);
+#endif  // INCLUDE_WAYS_H_
